Move "npm-install" task definition to plugins.js
